+++ src/krama/fhir/__init__.py
@@ -5,9 +5,11 @@ from krama.fhir.compositions import PrescriptionBuilder
 from krama.fhir.bundles import create_discharge_summary_bundle
 from krama.fhir.bundles import create_op_consult_bundle
 from krama.fhir.bundles import create_prescription_bundle
+from krama.fhir.facade import FHIRFacade
 from krama.fhir.resources import PatientIdentifier, PatientIdentifierType
 
 __all__ = [
+    "FHIRFacade",
     "OPConsultBuilder",
     "PatientIdentifier",
     "PatientIdentifierType",